Added ajax Battle form with hashtag input on main page and display tweet and the counter
To be able to pull the data from twitter streaming API, We need to add Event Listener in javascript. After user fill out hashtags and press submit, we will insert battle name and their hashtags into database and call twitter streaming API. We match hashtag input with those tweet. If it's match, we will publish tweet and keywords in battle.hashtag channel in JSON format. On client side, we have Event listener to fetch data and display with HTML template